requirements.txt
flask
gunicorn
requests
python-docx
pypdf2
python-dotenv
reportlab
